requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
google-genai
cryptography
pydantic